JFR: Setting graphics to High, Adding Post Processing Stack and getting the hands and controllers to show up again.
Set up the graphics to a pretty high setting to make sure that everythign looks as nice as possible. Adding Postprocessing stack and enabled a couple of effects. Also reimported the Oculus avatar and Camera rig to solve the problem of why the controllers and hands where not showing up where they should have been.